DOCKER_BACKEND/device_backend/requirementslocal.txt
alabaster==0.7.12
appdirs==1.4.4
argon2-cffi==20.1.0
asgiref==3.3.1
astroid==2.4.2
atomicwrites==1.4.0
attrs==20.3.0
Babel==2.9.0
backcall==0.2.0
black==20.8b1
certifi==2020.12.5
cffi==1.14.4
cfgv==3.2.0
chardet==4.0.0
click==7.1.2
colorama==0.4.4
coverage==5.3.1
cryptography==3.3.1
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.1
Django==3.0.11
django-allauth==0.44.0
django-coverage-plugin==1.8.0
django-crispy-forms==1.10.0
django-debug-toolbar==3.2
django-environ==0.4.5
django-extensions==3.1.0
django-model-utils==4.1.1
django-redis==4.12.1
django-stubs==1.7.0
djangorestframework==3.12.2
docutils==0.16
factory-boy==3.1.0
Faker==5.6.0
filelock==3.0.12
flake8==3.8.4
flake8-isort==4.0.0
identify==1.5.12
idna==2.10
imagesize==1.2.0
iniconfig==1.1.1
ipdb==0.13.4
ipython==7.19.0
ipython-genutils==0.2.0
isort==5.7.0
jedi==0.18.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
livereload==2.6.3
Markdown==3.3.3
MarkupSafe==1.1.1
mccabe==0.6.1
mypy==0.790
mypy-extensions==0.4.3
nodeenv==1.5.0
oauthlib==3.1.0
packaging==20.8
parso==0.8.1
pathspec==0.8.1
pickleshare==0.7.5
Pillow==8.1.0
pluggy==0.13.1
pre-commit==2.9.3
prompt-toolkit==3.0.10
psycopg2-binary==2.8.6
py==1.10.0
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
Pygments==2.7.4
PyJWT==2.0.0
pylint==2.6.0
pylint-django==2.4.2
pylint-plugin-utils==0.6
pyparsing==2.4.7
pytest==6.2.1
pytest-django==4.1.0
pytest-sugar==0.9.4
python-dateutil==2.8.1
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2020.5
PyYAML==5.3.1
redis==3.5.3
regex==2020.11.13
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
snowballstemmer==2.0.0
Sphinx==3.4.3
sphinx-autobuild==2020.9.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.4.1
termcolor==1.1.0
testfixtures==6.17.1
text-unidecode==1.3
toml==0.10.2
tornado==6.1
traitlets==5.0.5
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.2
virtualenv==20.3.1
wcwidth==0.2.5
websockets==8.1
Werkzeug==1.0.1
whitenoise==5.2.0
wrapt==1.12.1
